rename gpsModal to GPSPermissionModal also renamed other connected compontents
