+++ src/frontend/screens/MapScreen/GPSPermissions/GPSPermissionsEnabled.tsx
